add usbPortManager by hilt
diでの実装が完了したが、usbではなくBluetoothでの実装でないといけないことが発覚したので一旦コメントアウト
