+++ app/src/main/java/com/example/usb_host_tutorial/ui/viewmodel/MainViewModel.kt
@@ -11,5 +11,7 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(
 //    val usbPortManager: UsbPortManager
 ): ViewModel() {
-
+//    init {
+//        usbPortManager.openPort()
+//    }
 }
